feat: Done with landing page
